+++ src/netius/base/client.py
@@ -182,7 +182,9 @@ class DatagramClient(Client):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
@@ -218,7 +220,9 @@ class DatagramClient(Client):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
@@ -863,7 +867,9 @@ class StreamClient(Client):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
@@ -912,7 +918,9 @@ class StreamClient(Client):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS: