+++ src/netius/base/common.py
@@ -3053,7 +3053,9 @@ class AbstractBase(observer.Observable):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
@@ -3102,7 +3104,9 @@ class AbstractBase(observer.Observable):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
@@ -3147,7 +3151,9 @@ class AbstractBase(observer.Observable):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
@@ -3448,7 +3454,9 @@ class AbstractBase(observer.Observable):
             error_v = error.args[0] if error.args else None
             error_m = (
                 error.reason.upper().replace(" ", "_")
-                if hasattr(error, "reason") and error.reason
+                if not error_v in SSL_VALID_ERRORS
+                and hasattr(error, "reason")
+                and error.reason
                 else None
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS: